requirements.txt
dj-database-url==0.5.0
Django==2.0
django-environ==0.4.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
waitress==1.3.0
whitenoise==4.1.2
